% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bachelors}
\alias{bachelors}
\title{US Bachelors Degrees}
\format{A data frame containing thirty-three observations for nineteen years.}
\source{
U.S. Department of Education, National Center for Education Statistics, Higher Education General Information Survey (HEGIS), \dQuote{Degrees and Other Formal Awards Conferred} surveys, 1970-71 through 1985-86; Integrated Postsecondary Education Data System (IPEDS), \dQuote{Completions Survey} (IPEDS-C:91-99); and IPEDS Fall 2000 through Fall 2014, Completions component. Retrieved 2016-12-08 from \url{http://nces.ed.gov/programs/digest/d15/tables/dt15_322.10.asp}.
}
\usage{
bachelors
}
\description{
Bachelor's degrees conferred by postsecondary institutions, by field of study: Selected years, 1970-71 through 2013-14
}
\keyword{datasets}
