\name{cancer}
\docType{data}
\alias{cancer}
\title{Estimates of relative survival rates, by cancer site}
\description{Cancer survival rates for 5, 10, 15, and 20 years for twenty-four cancer sites.}
\usage{cancer}
\format{A dataframe containing twenty-four observations for four years.}
\source{Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.}
\keyword{datasets}
