% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer2}
\alias{cancer2}
\title{Estimates of relative survival rates, by cancer site (other)}
\format{A data frame containing long-formatted data from \code{\link{cancer}}, yielding 96 observations for variables \dQuote{Type}, \dQuote{Year}, and \dQuote{Survival} (surival rate).}
\source{
Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
}
\usage{
cancer2
}
\description{
Cancer survival rates for 5, 10, 15, and 20 years for twenty-four cancer sites.
}
\seealso{
For data in \dQuote{wide} format, see \code{\link{cancer}}
}
\keyword{datasets}
