\name{gdp}
\docType{data}
\alias{gdp}
\title{Current receipts of government as a percentage of Gross Domestic Product, 1970 and 1979}
\description{Current receipts of fifteen national governments as a percentage of gross domestic product}
\usage{gdp}
\format{A dataframe containing fifteen country observations for two years.}
\source{Edward Tufte. \emph{The Visual Display of Quantitative Information}. Graphics Press, 158-159.}
\keyword{datasets}