% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{segmentize}
\alias{segmentize}
\title{Segmentize an observation-by-period data frame}
\usage{
segmentize(data)
}
\arguments{
\item{data}{A data frame containing observation-by-period data, where the only columns represent sequentially ordered time period values for each observation. Some values can be missing.}
}
\value{
A five-column matrix containing: the row from the original data frame, and x1, x2, y1, y2 positions for each segment.
}
\description{
Convert an observation-by-period data frame into a matrix of line segment coordinates, where each row represents a line segment connecting two time points from the original data.
}
\examples{
data(gdp)
head(segmentize(gdp))
}

