\name{states}
\docType{data}
\alias{states}
\title{U.S. State Populations}
\description{The relative rank of U.S. state populations for nine decades (1790-1870).}
\usage{states}
\format{A dataframe containing twenty-four observations for four years.}
\source{Gray, Ormando Willis, and Colton, G. Woolworth. 1878. ``Chart Exhibiting the Relative Rank of the States for Nine Decades (1790-1870).'' Philadelphia: O.W. Gray & Son. Dataset constructed from slopegraph shown here: \url{http://www.davidrumsey.com/luna/servlet/detail/RUMSEY~8~1~207741~3003452:Chart-Exhibiting-the-Relative-Rank-}}
\keyword{datasets}
