% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Estimates of relative survival rates, by cancer site}
\format{A data frame containing twenty-four observations for four years.}
\source{
Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
}
\usage{
cancer
}
\description{
Cancer survival rates for 5, 10, 15, and 20 years for twenty-four cancer sites.
}
\keyword{datasets}
