% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Current receipts of government as a percentage of Gross Domestic Product, 1970 and 1979}
\format{A data frame containing fifteen country observations for two years.}
\source{
Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
}
\usage{
gdp
}
\description{
Current receipts of fifteen national governments as a percentage of gross domestic product
}
\keyword{datasets}
