% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggslopegraph.R
\name{ggslopegraph}
\alias{ggslopegraph}
\title{Create Slopegraph from a data frame using ggplot2}
\usage{
ggslopegraph(data, title = as.character(substitute(data)), xlab = "",
  ylab = "", yrev = FALSE, size = 3)
}
\arguments{
\item{data}{An observation-by-period data.frame, with at least two columns. Missing values are allowed.}

\item{title}{A character string specifying a title. Passed to \code{\link[ggplot2]{ggtitle}}.}

\item{xlab}{A character string specifying an x-axis label. Passed to \code{\link[ggplot2]{scale_x_continuous}}.}

\item{ylab}{A character string specifying an y-axis label. Passed to \code{\link[ggplot2]{scale_y_continuous}}.}

\item{yrev}{A logical indicating whether to use \code{\link[ggplot2]{scale_y_reverse}} rather than the default \code{\link[ggplot2]{scale_y_continuous}}.}

\item{size}{A numeric value indicating the size of numeric labels.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Convert an R data frame (containing a panel dataset, where rows are observations and columns are time periods) into an Edward Tufte-inspired Slopegraph using ggplot2
}
\examples{
## Ranking of U.S. State populations
data(states)
ggslopegraph(states, title = 'Relative Rank of U.S. State Populations, 1790-1870') + scale_y_reverse()
}
\seealso{
For a base graphics version, use \code{\link{slopegraph}}.
}

