% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize.R
\name{segmentize}
\alias{segmentize}
\title{Segmentize an observation-by-period data frame}
\usage{
segmentize(data, na.span = FALSE, na.omit = TRUE)
}
\arguments{
\item{data}{A data frame containing observation-by-period data, where the only columns represent sequentially ordered time period values for each observation. Some values can be missing.}

\item{na.span}{A logical indicating whether line segments should span periods with missing values. The default is \code{FALSE}, such that some segments are not drawn.}

\item{na.omit}{A logical indicating whether to drop missing observations from the resulting data frame}
}
\value{
A five-variable data frame containing: the row from the original data frame, and x1, x2, y1, y2 positions for each segment.
}
\description{
Convert an observation-by-period data frame into a data frame of line segment coordinates, where each row represents a line segment connecting two time points from the original data.
}
\examples{
data(gdp)
head(segmentize(gdp))
}
